feat: Remove unused files and components
This commit removes 33 unused files from the codebase, including UI components, hooks, stylesheets, and data files.

The following files were removed:
- 28 unused UI components from `src/shared/ui`
- 2 unused hooks from `src/shared/hooks`
- 1 unused stylesheet `src/App.css`
- 1 unused data file `src/martin-info.json`
- 1 unused types file `src/types.ts`

Additionally, this commit fixes a circular dependency between the `ui` and `hooks` layers by moving the `useToast` hook implementation to the `ui` directory. All related import paths and type definitions have been updated accordingly.
