+++ package.json
@@ -80,6 +80,7 @@
     "globals": "^15.9.0",
     "next": "^15.4.5",
     "tailwindcss": "^3.4.11",
+    "ts-prune": "^0.10.3",
     "typescript": "^5.4.5",
     "typescript-eslint": "^8.0.1"
   }
