--- src/shared/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from "./use-mobile";
-export * from "./use-toast";
