+++ src/shared/ui/index.ts
@@ -1,21 +1,9 @@
-export * from "./accordion";
-export * from "./alert-dialog";
-export * from "./alert";
-export * from "./aspect-ratio";
-export * from "./avatar";
 export * from "./badge";
-export * from "./breadcrumb";
 export * from "./button";
-export * from "./calendar";
 export * from "./card";
 export * from "./carousel";
-export * from "./chart";
 export * from "./checkbox";
-export * from "./collapsible";
-export * from "./command";
-export * from "./context-menu";
 export * from "./dialog";
-export * from "./drawer";
 export * from "./dropdown-menu";
 export * from "./form";
 export * from "./glass-card";
@@ -23,34 +11,18 @@ export * from "./glow-effect";
 export * from "./gradient-text";
 export * from "./hover-card";
 export * from "./image-carousel";
-export * from "./input-otp";
 export * from "./input";
 export * from "./label";
-export * from "./menubar";
-export * from "./navigation-menu";
-export * from "./pagination";
-export * from "./popover";
-export * from "./progress";
-export * from "./radio-group";
-export * from "./resizable";
 export * from "./scroll-area";
-export * from "./select";
 export * from "./separator";
-export * from "./sheet";
-export * from "./sidebar";
-export * from "./skeleton";
-export * from "./slider";
 export * from "./smart-image";
 export * from "./sonner";
 export * from "./switch";
-export * from "./table";
-export * from "./tabs";
 export * from "./textarea";
 export * from "./theme-toggle";
 export * from "./toast";
 export { Toaster as ToasterComponent } from "./toaster";
 export { toast as toastUtil } from "./use-toast";
-export * from "./toggle-group";
 export * from "./toggle";
 export * from "./tooltip";
 export * from "./badge-utils";
