--- src/shared/ui/skeleton.tsx
@@ -1,15 +1,0 @@
-import { cn } from "@/shared/lib"
-
-function Skeleton({
-  className,
-  ...props
-}: React.HTMLAttributes<HTMLDivElement>) {
-  return (
-    <div
-      className={cn("animate-pulse rounded-md bg-muted", className)}
-      {...props}
-    />
-  )
-}
-
-export { Skeleton }
